requirements.txt
influxdb~=5.3.1
RPi.GPIO~=0.7.0
adafruit-circuitpython-dht~=3.5.1